refactor(core): split epoch manager actor in different files
